backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
fastapi==0.115.4
filelock==3.18.0
Flask==2.3.3
Flask-Cors==4.0.0
fsspec==2025.3.0
google-auth==2.39.0
google-auth-oauthlib==1.2.2
google_search_results==2.4.2
gspread==6.2.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.29.3
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==7.4.0
python-dotenv==1.0.0
pytweening==1.2.0
pywhatkit==5.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
rubicon-objc==0.5.0
safetensors==0.5.3
selenium==4.31.0
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.41.2
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.49.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.32.0
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wsproto==1.2.0
